Address direct path repair review feedback
