+++ crates/mesh-llm-protocol/src/protocol/mod.rs
@@ -50,6 +50,7 @@ pub enum ControlFrameError {
     BadGeneration { got: u32 },
     InvalidEndpointId { got: usize },
     InvalidSenderId { got: usize },
+    MissingDirectPathAddress,
     MissingHttpPort,
     MissingOwnerId,
     MissingControlOwnerId,
@@ -88,6 +89,9 @@ impl std::fmt::Display for ControlFrameError {
             }
             ControlFrameError::InvalidSenderId { got } => {
                 write!(f, "invalid sender_id length: expected 32, got {}", got)
+            }
+            ControlFrameError::MissingDirectPathAddress => {
+                write!(f, "direct path request missing endpoint address")
             }
             ControlFrameError::MissingHttpPort => {
                 write!(f, "HOST-role peer annotation missing http_port")
@@ -257,9 +261,7 @@ impl ValidateControlFrame for crate::proto::node::DirectPathRequest {
             });
         }
         if self.serialized_addr.is_empty() {
-            return Err(ControlFrameError::DecodeError(
-                "direct path request missing endpoint address".to_string(),
-            ));
+            return Err(ControlFrameError::MissingDirectPathAddress);
         }
         Ok(())
     }
